+++ src/map/error.rs
@@ -8,11 +8,6 @@ pub enum Error {
   /// Indicates that the arena is full
   Full(ArenaError),
 
-  /// Indicates that an entry with the specified key already
-  /// exists in the skiplist. As a low-level crate, duplicate entries are not directly supported and
-  /// instead must be handled by the user.
-  Duplicated,
-
   /// Indicates that the value is too large to be stored in the [`SkipMap`](super::SkipMap).
   ValueTooLarge(u64),
 
@@ -30,7 +25,6 @@ impl core::fmt::Display for Error {
   fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
     match self {
       Self::Full(e) => write!(f, "{e}"),
-      Self::Duplicated => write!(f, "key already exists in the skipmap"),
       Self::ValueTooLarge(size) => write!(f, "value size {} is too large", size),
       Self::KeyTooLarge(size) => write!(f, "key size {} is too large", size),
       Self::EntryTooLarge(size) => write!(f, "entry size {size} is too large",),
@@ -51,10 +45,6 @@ impl From<ArenaError> for Error {
 #[cfg(test)]
 #[test]
 fn test_fmt() {
-  assert_eq!(
-    std::format!("{}", Error::Duplicated),
-    "key already exists in the skipmap"
-  );
   assert_eq!(
     std::format!("{}", Error::KeyTooLarge(10)),
     "key size 10 is too large"